requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
numpy==1.24.3
faiss-cpu==1.7.4
google-generativeai==0.3.2